+++ tests/server/blueprints/variant/test_variant_controllers.py
@@ -198,10 +198,9 @@ def test_observations_controller_non_existing(app, institute_obj, case_obj, loqu
     ## WHEN updating the case_id for the variant
     var_obj["case_id"] = "internal_id2"
 
-    data = None
     with app.test_client() as client:
-        resp = client.get(url_for("auto_login"))
-        data = observations(store, loqusdb, case_obj, var_obj)
+        client.get(url_for("auto_login"))
+        data = observations(store, loqusdb, var_obj)
 
     ## THEN assert that the number of cases is still returned
     assert data[loqus_id]["total"] == n_cases
@@ -241,10 +240,9 @@ def test_observations_controller_snv(app, institute_obj, loqusdburl):
     ## WHEN updating the case_id for the variant
     var_obj["case_id"] = "internal_id2"
 
-    data = None
     with app.test_client() as client:
-        resp = client.get(url_for("auto_login"))
-        data = observations(store, loqusdb, case_obj, var_obj)
+        client.get(url_for("auto_login"))
+        data = observations(store, loqusdb, var_obj)
 
     ## THEN loqus should return the occurrence from the first case
     assert case_obj["_id"] in data[loqus_id]["families"]
@@ -286,9 +284,9 @@ def test_observations_controller_sv(app, sv_variant_obj, institute_obj, loqusdbu
     sv_variant_obj["variant_id"] = "someOtherVarID"
 
     with app.test_client() as client:
-        resp = client.get(url_for("auto_login"))
+        client.get(url_for("auto_login"))
         # THEN the observation of the original case should be found
-        data = observations(store, loqusdb, case_obj, sv_variant_obj)
+        data = observations(store, loqusdb, sv_variant_obj)
 
     ## THEN loqus should return the occurrence from the first case
     assert case_obj["_id"] in data[loqus_id]["families"]