+++ CHANGELOG.md
@@ -11,6 +11,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Revel score, Revel rank score and SpliceAI values are also displayed in Causatives and Validated variants tables
 - Remove unused functions and tests
 - Analysis type and direct link from cases list for OGM cases
+- Removed unused `case_obj` parameter from server/blueprints/variant/controllers/observations function
 - Replace the code that downloads Ensembl genes, transcripts and exons with schug library
 ### Fixed
 - Disease_term identifiers are now prefixed with the name of the coding system